Don't ignore custom sample mime types when resolving track type
When the sample mime type resolves to UNKNOWN, we currently fall back
to the container mime type to resolve the track type. This is unexpected
for sample mime types that are set, but not recognized, because the
container mime type is often set to the "primary" track in a muxed file.
We should only use the container mime type if the sample mime type is
not set at all.

Issue: androidx/media#2860

#cherrypick

PiperOrigin-RevId: 834842598