Refactor button states to use shared PlayerStateObserver
All button state have a repeating pattern of updating their internal
state values in 3 different places. The observe method is also
re-implemented in the same pattern in all classes. This allows us
to move the shared pattern to a helper class, so no state update
code duplication is needed and all relevant logic is in a single
readable place.

PiperOrigin-RevId: 834853446