Fix error acumulation when converting timestamps to durations
In `Boxes#convertPresentationTimestampsToDurationsVu()`, we were
accumulating errors when calculating the duration of each timestamp in
video units. Each duration was being calculated as the delta of two
timestamps in microseconds, and then converted to video units. Each
duration was incurring on a small truncation loss that accumulated over
time to multiple ticks, even if each calculation was within the allowed
tolerance.

Instead, the method should first convert each timestamp to video units
and then calculate the difference between the two. This way, any error
can self correct because the source of truth for the conversion is the
timestamp itself.

With this fix, any timestamps converted by
`convertPresentationTimestampsToDurationsVu()` should map back to within
half a tick of their original values.

PiperOrigin-RevId: 835184938