Extract metrics collection setup to a helper method
The logic for setting up the `EditingMetricsCollector` and obtaining the `LogSessionId` is moved from `startInternal()` to a new private method `setUpMetricsCollection()`.

This new helper method is also called from `cancel()` and `onExportCompletedWithError()` to ensure that metrics are collected even in cancellation and error scenarios.

This is pre-work for 2 issues: concurrent exports in Transformer, and extracting track types for single-asset exports. Calling this new helper method in `onExportCompletedWithError()` ensures `EditingMetricsCollector` is set up even if Transformer is throwing before `startInternal()` is called.

PiperOrigin-RevId: 835229764