Deflake Scrubbing test for skip keyframe reset with b-frame content
`ScrubbingPlaybackTest::scrubbingPlayback_withSkipKeyFrameResetEnabledAndNonSequentialFrames_dumpsCorrectOutput` was flaky because depending on the "current" playback position when enabling scrubbing mode, the sync-point may differ between the seek and current playback position. By advancing the player until the playback position is past keyframe of the content then the sync points will be the same and scrubbing mode will seek without the keyframe reset.

PiperOrigin-RevId: 836148821