Fix clearing of state between instrumentation test runs on gradle
The previous config was:
1. Only applied in 5 modules
2. Not working, because it only applied 1 of the 3 steps described in
   https://developer.android.com/training/testing/instrumented-tests/androidx-test-libraries/runner

This change moves the config to `common_library_config.gradle` so that
it is inherited by all library configs.

This change was tested by showing that the following command fails
before it due to a hang in
`DefaultAudioSinkTest.audioTrackExceedsSharedMemory_retriesUntilOngoingReleasesAreDone`
(despite this test passing when run in isolation from Gradle):

```
./gradlew :lib-exoplayer:connectedDebugAndroidTest
```

The same command passes with this change, because whatever cross-test
pollution was causing the failure is resolved by the data clearing.

PiperOrigin-RevId: 836641698