Fix Protobuf gradle setup
The current setup fails to run from the command line in some
cases before the protobuf plugin is applied too early.

The correct setup is to declare the plugin as a dependency in
the high-level build.gradle file and then simply load the
plugin after initializing the shared module settings. This
matches how the Kotlin Gradle plugin is set up as well.

PiperOrigin-RevId: 837458389