Deflake ExoPlayerScrubbing seekAfterDisablingScrubbingMode unit test
The unit test `scrubbingMode_seekAfterDisablingScrubbingMode_preemptsActiveSeek` tests that a seek after disabling scrubbing mode will not be queued and will pre-empt any actively processing seek. The test was a little flaky because sometimes the player would advance and process a few frames before the `VideoFrameMetadataListener` was cleared. The solution was to run in a paused state so that playback won't continue after seeks are completed. The point of the test of pre-empting active seeks is still tested.

PiperOrigin-RevId: 837462290
(cherry picked from commit 44fb298b1d4b9eca18e3c64890ebf52be174753c)