Add util to transfer byte arrays over Bundles
Byte arrays that are not guaranteed to be very small shouldn't
be directly transferred via Bundle to remote processes because
there is a risk they exceed the inherent binder size limits.

Instead, such data should be transferred via separate shared
memory spaces (this technique is also used to transfer Bitmap
instances for example). On API 23-26 we need a workaround to
split up the data if it's too large and transfer it in chunks
instead.

PiperOrigin-RevId: 838709973