Fix flaky video frame processor flush test
There is still a race condition between `onTextureRendered` outputting the second
frame from the first item, and the test runner calling `flush()`

This test is only intended to check that all the frames output to the listener
after the listener receives `flush()` are from after the flush. The earlier
assertion was too strong.

PiperOrigin-RevId: 839219670