Create the ImaAdTagUriBuilder class
Publishers will be able to use this class to set their preferred VideoOrientation for ads.
We will use this value in serving to filter Mediafiles based on orientation, and in the JS Core to select a MediaFile from the resulting VAST.

Support video orientation controls using the ImaAdTagUriBuilder.
This feature is currently being alpha tested, if you would like to be added to the alpha testing group please reach out to your account partner.

PiperOrigin-RevId: 839219715