Fix issue where timeline was being set prior to disable and enable
In `handleMediaSourceListInfoRefreshed`, `Renderer.setTimeline` is invoked with the new timeline. After setting the timeline, depending on what what changed with the media, `handleMediaSourceListInfoRefreshed` may also incur a seek that disables and re-enables the renderers. In this case the `Renderer.setTimeline` operation should occur between the disable and enable operations. A `Renderer` should not be called with `setTimeline` if it currently enabled on a old, invalid `MediaPeriod`. Likewise a Renderer should not be enabled on a `MediaPeriod` while it has an old, invalid `Timeline`.

PiperOrigin-RevId: 839279241