Deflake OggOpusPlaybackTest
The test asserts the bytes written to the AudioTrack (containing the
OPUS samples wrapped in a OGG container). It's currently flaky
because we write the samples in opportunistic batches depending on
how quickly we can read from the source. For actual playback, it
doesn't matter where the batches are split though for as long as the
overall byte structure is correct. Changing the dump the assert the
overall sent bytes are correct prevents the flakiness.

PiperOrigin-RevId: 839281904