Move silence generation logic into an AudioProcessor
This change decouples silence appending logic from `AudioGraphInput`
into a standalone `AudioProcessor` that appends silence if a stream
does not reach a specified input duration.

This is prework for adding a "preprocessing" pipeline in
`AudioGraphInput` for non-user set effects that need to be applied
before user-set effects. This is necessary for handling speed
changing or format conversions.

PiperOrigin-RevId: 839813409