Fix IndexOutOfBoundsException in unclipped mode
`AdsMediaSource` allows a state where the
`AdsMediaSourceHolder` already contains a holder for
a source when the corresponding ad count in the
`AdPlaybackState` is still zero. This can happen
when the source gets a request to create a
`MediaPeriod` and the `AdPlaybackState` only contains
the position of the ad group yet.

Added a unit test to assert this corner case is handled
correctly.

PiperOrigin-RevId: 840233883