Fix flakes in CompositionPlayerAudioPlaybackTest dump file tests
`CompositionPlayerAudioPlaybackTest` uses a `CapturingAudioSink` to
record the output audio fed to the final `AudioSink`. The current tests
are too strict because they assert on the number of buffers that the
AudioGraph produces, which is dependent on the timing of the
`AudioGraph#getOutput()` calls. The new test implementation asserts
directly on the full output produced by `CompositionPlayer`.

This change introduces a new `CapturingAudioSink` configuration that
records all samples input into the sink independently of how they are
batched into individual buffers. This is also relevant for other tests
that want to assert on output, but not necessarily on how the buffer
structure (e.g. `OggOpusPlaybackTest`).

PiperOrigin-RevId: 840726805