Update session demo to use standard command buttons
It's currently redefining a standard command as a custom one, which we
explicitly advise against in documentation. Doing this also removes
the need to use the onCustomCommand override, which doesn't need to
be shown in the demo app as almost all button customizations should
work by using standard commands anyway.

Also add TODO for existing/new pattern of using the Unconfined scope
for listener registration. It should use the service lifecycle instead
to show best practices, even though our demo service is not stopped
before the entire process is destroyed anyway.

#cherrypick

PiperOrigin-RevId: 841725791