Fix flakiness in EndToEndOffloadFailureRecoveryTest
The test was flaky for same reason why https://github.com/androidx/media/commit/1beb1b4fbdf6cefe3a570e9d54aa8a67123e81fc was
needed: the batched OggOpus buffers can't be asserted
individually as they depend on reading/loading speed.
Instead we use the same idea here and assert the overall
written bytes are the same.

For the write error specifically, this means we need to
limit the buffer at a specific byte position as well
(and not at "number of samples") to get a reliable
behavior.

PiperOrigin-RevId: 842144777