[ui-compose-material3] Add progress indicator/slider
To create timebar functionality, we need to unite the Material3 `Slider` Composable and `ProgressIndicator` Composable. One gives the user ability to pick a value by scrubbing the `Thumb`. Another is driven by a state-backed value to move the `Thumb` along the `Track` automatically. There is currently no out-of-the-box solution for it in Material3. Sliders are used for scrub-and-done kind of action (e.g. volume, brightness) and progress indicators are used to show a non-interactive bar (e.g. download progress). Playback case is special and needs to both display the progress of the Player, but also let the user seek within the `MediaItem`.

Note that the current design does not allow for displaying of "buffering position" as a different colour on the track.

See:
* https://m3.material.io/components/sliders
* https://m3.material.io/components/progress-indicators

http://Issue: androidx/media#2288

#cherrypick

PiperOrigin-RevId: 843677725