Increase limit for acceptable audio latency values
Some output pipelines in the Audio HAL may take longer than
5 seconds. Since this is just a fail-safe, we can increase it
slightly to allow these larger values while still filtering
out impossibly large values that far exceed the threshold.

In addition, we can force a new getTimestamp call if the
return latency value changes, as this implies a change in
the output processing pipeline, which could mean the
interpolation of the previous timestamp may not be fully
reliable anymore.

PiperOrigin-RevId: 845760137