Don't use static session map for session look up
The lookup by session URI in `onStartCommand` was done in the static
map of sessions that is scoped by the `MediaSession` class. This can
have the consequence that a first `Intent` that is sent to the service
looks up a session that hasn't been added to the service yet.

Adding a lookup method that is scoped to the service instance is much
clearer. Some very special (aka. weird) use cases are not possible
anymore after this change. Before this change and when running in a
single process a session can be created for instance in an `Activity`
before the service even started to then start the service by `Intent`
to look up that very session by URI.

It's unclear if it ever was intended to provide this use case given it
only works when run in the same process. Removing this feature is
justified by having a less confusing approach that works the same when
running in a single or multiple processes.

It also prevents the service from relying on a static map that is in
itself questionable. The only feature that is based on this static map
should be to ensure that an app can never create two session with the
same ID within the same process. See the constructor of `MediaSession`
where this is still ensured

PiperOrigin-RevId: 846753110