Remove hardcoded SDK 29 checks in MCAR and DefaultAudioSink
There are some device specific APIs for offload on earlier APIs.
Removing these checks allows these device specific APIs to be used
by customizing AudioOutputProvider and OffloadSupportProvider,
without having to change MCAR or DefaultAudioSink.

For AudioTrackAudioOutputProvider and DefaultOffloadSupportProvider,
this change is a no-op. They check for SDK 29 on their own and will
just not do anything related to offload on earlier SDKs.